requirements/all.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.1
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
jwcrypto==1.0
matplotlib-inline==0.1.3
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.0
